requirements.txt
absl-py==0.10.0
aiohttp==3.6.2
asgiref==3.2.10
astunparse==1.6.3
async-timeout==3.0.1
attrs==20.2.0
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
cycler==0.10.0
decorator==4.4.2
Django==3.1.1
gast==0.3.3
google-auth==1.22.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.32.0
gunicorn==20.0.4
h5py==2.10.0
idna==2.10
imageio==2.9.0
importlib-metadata==2.0.0
joblib==0.16.0
Keras==2.4.3
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
Markdown==3.2.2
matplotlib==3.3.2
multidict==4.7.6
networkx==2.5
numpy==1.18.5
oauthlib==3.1.0
opencv-python==4.4.0.44
opt-einsum==3.3.0
pandas==1.1.2
Pillow==7.2.0
protobuf==3.13.0
pyasn1==0.4.8
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
PyWavelets==1.1.1
PyYAML==5.3.1
rando==0.1.3
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
scikit-learn==0.23.2
scipy==1.5.2
six==1.15.0
sqlparse==0.3.1
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.1
tensorflow-estimator==2.3.0
termcolor==1.1.0
threadpoolctl==2.1.0
tifffile==2020.9.28
tqdm==4.50.0
typing-extensions==3.7.4.3
urllib3==1.25.10
Werkzeug==1.0.1
whitenoise==5.2.0
wrapt==1.12.1
yarl==1.6.0
